requirements.txt
fal-client>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
pydantic>=2.0.0
volcengine-python-sdk[ark]>=1.0.0
